+++ packages/recommend/src/types/RelatedProductsQuery.ts
@@ -1,0 +1,3 @@
+import { RecommendationsQuery } from './RecommendationsQuery';
+
+export type RelatedProductsQuery = Omit<RecommendationsQuery, 'model'>;