+++ rollup.config.js
@@ -29,6 +29,7 @@ const packagesConfig = [
   'client-account',
   'client-analytics',
   'client-common',
+  'client-personalization',
   'client-recommendation',
   'logger-common',
   'logger-console',
@@ -80,6 +81,7 @@ packagesConfig.push({
   packagesConfig.push({
     output: build === 'browser' ? 'algoliasearch' : 'algoliasearch-lite',
     package: 'algoliasearch',
+    name: 'algoliasearch',
     input: `src/builds/${build}.ts`,
     formats: ['esm-browser'],
     ignore: ['crypto'],
@@ -88,6 +90,7 @@ packagesConfig.push({
   packagesConfig.push({
     output: build === 'browser' ? 'algoliasearch' : 'algoliasearch-lite',
     package: 'algoliasearch',
+    name: 'algoliasearch',
     input: `src/builds/${build}.ts`,
     formats: ['umd'],
     external: ['dom'],
@@ -97,6 +100,23 @@ packagesConfig.push({
     ignore: ['crypto'],
   });
 });
+
+packagesConfig.push(
+  {
+    output: 'recommend',
+    package: 'recommend',
+    name: '@algolia/recommend',
+    input: `src/builds/browser.ts`,
+    formats: ['esm-browser', 'umd'],
+  },
+  {
+    output: 'recommend',
+    package: 'recommend',
+    name: '@algolia/recommend',
+    input: `src/builds/node.ts`,
+    formats: ['cjs'],
+  }
+);
 
 const packagesDir = path.resolve(__dirname, 'packages');
 const aliasOptions = { resolve: ['.ts'] };
@@ -143,7 +163,7 @@ packagesConfig
       const isEsmBrowserBuild = /\.esm.browser.js$/.test(output.file);
 
       if (isUmdBuild) {
-        output.name = 'algoliasearch';
+        output.name = packageConfig.name;
         output.banner = createLicence(output.file);
       }